service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
aiofiles==23.2.1
pydantic==2.5.0
torch>=1.9.0
torchvision>=0.10.0
open-clip-torch>=2.0.0
faiss-cpu>=1.7.0
Pillow>=8.3.0
numpy>=1.21.0
